+++ resources/activateCallback.js
@@ -2,7 +2,7 @@
 
 const {ceil} = Math;
 const arrayChunk = (arr, size) => {
-  return [...new Array(ceil(arr.length / size))].map((_, i) => {
+  return [...Array.from({length: ceil(arr.length / size)})].map((_, i) => {
     const offset = i * size;
     return arr.slice(offset, offset + size);
   });